Рефакторинг BurgerConstructorUnlockList и DraggableIngredient.
